feat: OD46S-101, OD46S-102, OD46S-103

+++ dainf-lab-server/src/main/java/br/edu/utfpr/dainf/model/ReservationItem.java
@@ -22,7 +22,7 @@ public class ReservationItem {
     private Long id;
 
     @Column(name = "quantity", nullable = false)
-    private Double quantity;
+    private BigDecimal quantity;
 
     @Column(name = "price", nullable = false)
     private BigDecimal price;
